requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# LLM Providers
openai>=1.0.0
zhipuai>=2.0.0
dashscope>=1.14.0  # 阿里云通义千问

# A股数据
akshare>=1.12.0
yfinance>=0.2.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
pypinyin>=0.51.0

# 前端
streamlit>=1.30.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# 情感分析
transformers>=4.35.0
torch>=2.0.0

# 向量存储与RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# 知识图谱
networkx>=3.0

# 可视化
plotly>=5.18.0
pyecharts>=2.0.0

# 日志
loguru>=0.7.0

# 异步支持
aiohttp>=3.9.0
